# src/lib/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cosmic Index API",
    "version": "1.0.0",
    "description": "A retrofuturistic web encyclopedia API for exploring cosmic objects. Aggregates data from NASA's Exoplanet Archive, JPL's Small-Body Database, NASA DONKI, CNEOS, and other NASA APIs into a unified read-only interface.",
    "contact": {
      "name": "Cosmic Index"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Cosmic Index API v1"
    },
    {
      "url": "/api",
      "description": "Cosmic Index User API"
    }
  ],
...
  "paths": {
...
    "/webhooks/stripe": {
      "post": {
        "operationId": "handleStripeWebhook",
        "summary": "Process Stripe webhook events",
        "description": "Verify the Stripe signature, apply subscription state changes, and record idempotency markers for duplicate suppression.",
        "tags": [
          "Billing"
        ],
        "responses": {
          "200": {
            "description": "Webhook accepted"
          },
          "400": {
            "description": "Missing or invalid Stripe signature"
          },
          "500": {
            "description": "Database, Stripe secret, or event processing failure"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ObservatoryCronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal bearer token configured through OBSERVATORY_CRON_SECRET."
      }
    },
...
  }
}